made changed for posts controller and controller file structure
